refactor wallet auth into a seperate file
